+++ examples/playwright/src/theia-toggle-bottom-indicator.ts
@@ -16,10 +16,6 @@
 
 import { TheiaStatusIndicator } from './theia-status-indicator';
 
-const TOGGLE_BOTTOM_ICON = 'codicon-window';
-
 export class TheiaToggleBottomIndicator extends TheiaStatusIndicator {
-    override async isVisible(): Promise<boolean> {
-        return super.isVisible(TOGGLE_BOTTOM_ICON);
-    }
+    id = 'bottom-panel-toggle';
 }