+++ packages/plugin-ext/src/common/plugin-api-rpc-model.ts
@@ -16,6 +16,7 @@
 
 import * as theia from '@theia/plugin';
 import type * as monaco from '@theia/monaco-editor-core';
+import { MarkdownString as MarkdownStringDTO } from '@theia/core/lib/common/markdown-rendering';
 import { UriComponents } from './uri-components';
 import { CompletionItemTag } from '../plugin/types-impl';
 import { Event as TheiaEvent } from '@theia/core/lib/common/event';
@@ -71,13 +72,7 @@ export interface Range {
     readonly endColumn: number;
 }
 
-export interface MarkdownString {
-    value: string;
-    isTrusted?: boolean;
-    uris?: {
-        [href: string]: UriComponents;
-    };
-}
+export { MarkdownStringDTO as MarkdownString };
 
 export interface SerializedDocumentFilter {
     $serialized: true;
@@ -107,7 +102,7 @@ export interface Completion {
     label2?: string;
     kind: CompletionItemKind;
     detail?: string;
-    documentation?: string | MarkdownString;
+    documentation?: string | MarkdownStringDTO;
     sortText?: string;
     filterText?: string;
     preselect?: boolean;
@@ -228,12 +223,12 @@ export enum MarkerTag {
 
 export interface ParameterInformation {
     label: string | [number, number];
-    documentation?: string | MarkdownString;
+    documentation?: string | MarkdownStringDTO;
 }
 
 export interface SignatureInformation {
     label: string;
-    documentation?: string | MarkdownString;
+    documentation?: string | MarkdownStringDTO;
     parameters: ParameterInformation[];
 }
 
@@ -251,7 +246,7 @@ export interface SignatureHelpContext {
 }
 
 export interface Hover {
-    contents: MarkdownString[];
+    contents: MarkdownStringDTO[];
     range?: Range;
 }
 
@@ -602,7 +597,7 @@ export enum CommentMode {
 
 export interface Comment {
     readonly uniqueIdInThread: number;
-    readonly body: MarkdownString;
+    readonly body: MarkdownStringDTO;
     readonly userName: string;
     readonly userIconPath?: string;
     readonly contextValue?: string;